removed all IF statements

+++ mux.sh
@@ -25,41 +25,24 @@ for i in *."$FILETYPE" ; do
 
   [ "$VCODEC" = "HEVC" ] && VSETTINGS="-c:v copy -map_metadata -1 -bsf:v hevc_mp4toannexb" && VFILE=0v-$RND.265
 
-  if [ "$ACODEC" = "AAC" ] ;
-  then
-    AFILE=0a-$RND.aac
-  elif [ "$ACODEC" = "E-AC-3" ] ;
-  then
-    ASETTINGS="-map 0:$ASTREAM -c:a ac3 -b:a $ABITRATE -map_metadata -1"
-  elif [ "$ACODEC" != "AC-3" ] ;
-  then
-    ASETTINGS="-map 0:$ASTREAM -c:a ac3 -b:a 640k -map_metadata -1"
-  fi
+  [ "$ACODEC" = "AAC" ] && AFILE=0a-$RND.aac
+  [ "$ACODEC" = "E-AC-3" ] && ASETTINGS="-map 0:$ASTREAM -c:a ac3 -b:a $ABITRATE -map_metadata -1"
+  [ "$ACODEC" != "AC-3" ] && ASETTINGS="-map 0:$ASTREAM -c:a ac3 -b:a 640k -map_metadata -1"
 
   [ -n "$SSTREAM" ] && SUBS="-map 0:$SSTREAM -c:s copy $FILENAME.srt"
 
   ffmpeg -i $i $VSETTINGS "$WORKINGDIR"/$VFILE $ASETTINGS "$WORKINGDIR"/$AFILE $SUBS -y -hide_banner ;
   rc=$?
 
-  if [ $rc -ne 0 ] ;
-  then
-    echo ERROR on ffmpeg : "$i" ;
-    exit 1
-  elif [ $DELETE -eq 1 ] ;
-  then
-    rm "$i" 2>/dev/null ;
-  fi
+  [ $rc -ne 0 ] && echo ERROR on ffmpeg : "$i" && exit 1
+  [ $DELETE -eq 1 ] && rm "$i" 2>/dev/null ;
 
   MP4Box -add "$WORKINGDIR/$VFILE#trackID=1:fps=$FRAMERATE:name=" -packed \
          -add "$WORKINGDIR/$AFILE#trackID=1:name=" -tmp "$WORKINGDIR" -new "$FILENAME.mp4" ;
   rc=$?
 
-  if [ $rc -ne 0 ] ;
-  then
-    echo ERROR on MP4Box : "$i" ;
-    exit 2
-  fi
+  [ $rc -eq 0 ] && rm "$WORKINGDIR"/$AFILE "$WORKINGDIR"/$VFILE 2>/dev/null || echo ERROR on MP4Box : "$i" && exit 2
+
 done
 
-rm "$WORKINGDIR"/$AFILE "$WORKINGDIR"/$VFILE 2>/dev/null ;
 exit 0
